Code clean up and fixes
* Re-implement move constructors in terms of move-assignment and remove various private copy constructors
* Add destructors for Gfx::Texture and Gfx::TextureSampler
* Add debug labels to Gfx::TextureSampler and make class movable
* Re-implement PosixFileIO using std::variant
* Make cast to IO* in PosixFileIO and GenericIO explicit to avoid unwanted implicit conversion to bool
* Remove various memory allocation wrappers
* Unset OpenGL sampler object before swapping buffers to fix blank display on Android SDK emulator
* Dynamically allocate PixmapTexture in GlyphEntry and use 16-bit ints for GlyphMetrics to save memory
* Only attempt to lock texture in initTextureCommon() if image source doesn't provide a locked pixmap
* Improve error handling in PBO case of Texture::lock() and unset GL_PIXEL_UNPACK_BUFFER binding so other texture functions aren't affected between call to unlock()
* Call glReleaseShaderCompiler() before entering background if its timer is scheduled
* Use eglInitialize()/eglTerminate() on threads with a GLDisplay and implement ref-counting in EGLContextBase to simulate Android behavior
* Use GLContext::bindAPI() directly after thread gets a GLDisplay

+++ imagine/include/imagine/font/GlyphMetrics.hh
@@ -1,15 +1,17 @@
 #pragma once
 
+#include <cstdint>
+
 namespace IG
 {
 
 struct GlyphMetrics
 {
-	int xSize = 0;
-	int ySize = 0;
-	int xOffset = 0;
-	int yOffset = 0;
-	int xAdvance = 0;
+	int16_t xSize = 0;
+	int16_t ySize = 0;
+	int16_t xOffset = 0;
+	int16_t yOffset = 0;
+	int16_t xAdvance = 0;
 
 	constexpr GlyphMetrics() {}
 };
